Started factoring out code into an explorer base and a slightly dumb random strategy.
