+++ comp313p_explorer/src/comp313p_explorer/explorer_node.py
@@ -1,16 +1,8 @@
 import rospy
-from comp313p_mapper.msg import *
 
-class ExplorerNode(object):
+from explorer_node_base import ExplorerNodeBase
 
-    def __init__(self):
-        rospy.init_node('explorer')
+class ExplorerNode(ExplorerNodeBase):
 
-        # Subscribe to get the map update messages
-        self.mapUpdateSubscriber = rospy.Subscriber('updated_map', MapUpdate, self.mapUpdateCallback)
 
-    def mapUpdateCallback(self, msg):
-        rospy.loginfo("map update received")
-    
-    def run(self):
-        rospy.spin()
+    pass
